Finished Assignment 3. Implemented interactive change of material properties and light color.

+++ AlmostGLVisualizer/AlmostGLVisualizer/source/Material.h
@@ -17,7 +17,7 @@ public:
     Vector3f ambient;
     Vector3f diffuse;
     Vector3f specular;
-    int shine;
+    float shine;
 };
